requirements.txt
flask
flask-socketio
python-socketio
python-dotenv
eventlet
bcrypt
customtkinter